Specify cache names
Without names, if the 2 methods with the same parameters are @Cacheable, then they'll use the same cache resulting in unexpected collisions. For example, invoke method A with the same parameter as method B had previously been invoked, and get back method B's return value instead of A's as one would expect.